+++ changelog.md
@@ -26,6 +26,8 @@
   it would be nice to give the user the control of choosing which sorting
   comes first.
 
+# 2022 04 04
+- Add menu option to client table row dropdown to create new invoice for client
 # 2022 04 03
 - Fix missing page after filtering invoices
 - add missing payto in invoice print view